Yolo/pyproject.toml
[project]
name = "license-plate-yolo-trainer"
version = "0.1.0"
description = "Minimal YOLO annotation and training app with inference APIs"
authors = [{name = "Codex", email = "codex@example.com"}]
requires-python = ">=3.10,<3.13"
dependencies = [
  "fastapi>=0.116.0,<0.117.0",
  "uvicorn[standard]>=0.35.0,<0.36.0",
  "jinja2>=3.1.6,<4.0.0",
  "python-multipart>=0.0.20,<0.0.21",
  "ultralytics>=8.3.0,<9.0.0",
  "pillow>=11.3.0,<12.0.0",
  "pyyaml>=6.0.2,<7.0.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
